requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.7
dnspython==2.3.0
filelock==3.10.4
Flask==2.2.3
frozenlist==1.3.3
greenlet==2.0.2
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.123
llama-index==0.4.38
loguru==0.6.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.24.2
openai==0.27.2
packaging==23.0
pandas==1.5.3
pinecone-client==2.2.1
pycodestyle==2.10.0
pydantic==1.10.7
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.2
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.47
tenacity==8.2.2
tokenizers==0.13.2
tomli==2.0.1
tqdm==4.65.0
transformers==4.27.3
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
